Fixed VC++ and GCC compiler warnings.

+++ AHN.Buildings/ClusterFilter.h
@@ -21,7 +21,7 @@ public:
 	/// <remarks>
 	/// 400 pixels is 100m2 with 0.5m AHN raster grid.
 	/// </remarks>
-	double sizeThreshold = 400;	
+	int sizeThreshold = 400;	
 	/// <summary>
 	/// <c>true</c> to test connectedness diagonally, <c>false</c> otherwise.
 	/// </summary>
